requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==42.0.8
Django==4.2.13
djangorestframework==3.15.2
numpy==1.24.4
opencv-python==4.10.0.84
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.0
typing-extensions==4.12.2
